Baal-main/iot/src/requirements.txt
adafruit-io
pyserial
firebase-admin
